修复数据分析出现的异常问题及添加特殊服务“其他”

+++ app/src/main/java/myapp/alex/com/businessassistant/activity/CreateOrderActivity.java
@@ -168,7 +168,9 @@ public class CreateOrderActivity extends AppCompatActivity implements CreateOrde
         list.add(orderServiceModel);
         orderAdapter.notifyDataSetChanged();
 
-        total = total + (model.getPrice() * Integer.parseInt(num));
+//        total = total + (model.getPrice() * Integer.parseInt(num));
+        //数量修改为float形式
+        total = total + (model.getPrice() *Float.parseFloat(num));
         money.setText("总金额:" + total + "元");
 
     }
